Don't bother recording at night to save more power
This change means that recordings stop between 10pm and 8am, according to the RTC - which I set to UTC. This definition of "night time" works for me in the UK where we're either on UTC time (GMT) or +1 (BST) and it makes little difference to me if "night time" is 10pm-8am at some times in the year or 11pm-9am in other times. If you lived in a different time zone then you might want to adjust this logic to sleep during more appropriate hours. Alternatively, you could keep this logic but set the RTC to be your local time and then work out whether any daylight saving hours changes are required based upon the date when you perform any post-processing yourself (if you set the RTC to your local time and your location has daylight savings changes then the time on the clock will be "wrong" for six months out of the year, which is why I decided to set it to UTC - it will always be correct in terms of UTC but I'll need to work out if the recorded time needs to have an hour added to it in Summer Time based upon the recorded date).